tweaking

+++ gobgp/cmd/global.go
@@ -783,7 +783,7 @@ func ParsePath(rf bgp.RouteFamily, args []string) (*table.Path, error) {
 		nlri, extcomms, err = ParseFlowSpecArgs(rf, args, rd)
 	case bgp.RF_DNS:
 		m := extractReserved(args, []string{"recordType", "key", "value"})
-		nlri = bgp.NewDnsNLRI(uint16(m["recordType"], m["key"], m["value"]))
+		nlri = bgp.NewDnsNLRI(m["recordType"], m["key"], m["value"]))
 	case bgp.RF_OPAQUE:
 		m := extractReserved(args, []string{"key", "value"})
 		if len(m["key"]) != 1 {
